OHRM5X-13:Updte job title with grid and delete confirmation

+++ symfony/client/src/orangehrmAdminPlugin/pages/jobTitle/JobTitle.vue
@@ -62,12 +62,15 @@
         />
       </div>
     </div>
+
+    <delete-confirmation ref="deleteDialog"></delete-confirmation>
   </div>
 </template>
 
 <script>
 import usePaginate from '@orangehrm/core/util/composable/usePaginate';
 import {navigate} from '@orangehrm/core/util/helper/navigation';
+import DeleteConfirmationDialog from '@orangehrm/components/dialogs/DeleteConfirmationDialog.vue';
 
 export default {
   data() {
@@ -100,6 +103,10 @@ export default {
       editItem: null,
       checkedItems: [],
     };
+  },
+
+  components: {
+    'delete-confirmation': DeleteConfirmationDialog,
   },
 
   setup() {
@@ -137,11 +144,18 @@ export default {
       this.checkedItems.forEach(index => {
         ids.push(this.items?.data[index].id);
       });
-      this.deleteItems(ids);
+      this.$refs.deleteDialog.showDialog().then(confirmation => {
+        if (confirmation === 'ok') {
+          this.deleteItems(ids);
+        }
+      });
     },
     onClickDelete(item) {
-      const id = item.id;
-      this.deleteItems([id]);
+      this.$refs.deleteDialog.showDialog().then(confirmation => {
+        if (confirmation === 'ok') {
+          this.deleteItems([item.id]);
+        }
+      });
     },
     deleteItems(items) {
       // TODO: Loading
